+++ src/core/Guild.cpp
