+++ src/core/Guild.hpp
